requirements.txt
amqp==2.5.2
billiard==3.6.3.0
celery==4.4.0
Click==7.0
Flask==1.1.1
Flask-CeleryExt==0.3.4
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
kombu==4.6.7
MarkupSafe==1.1.1
python-dotenv==0.12.0
pytz==2019.3
redis==3.4.1
six==1.14.0
SQLAlchemy==1.3.13
vine==1.3.0
Werkzeug==1.0.0
zipp==3.0.0
